extract translation fn, math.floor on canvas size
