+++ src/lib/scripts/layer.ts
@@ -73,3 +73,23 @@ export function createLayer(type: 'canvas' | 'text', name: string): Layer {
         };
     }
 }
+
+export function translateLayerBy(layer: Layer, dx: number, dy: number) {
+    const mat = layer.transform.matrix.translate(0, 0);
+    mat.m41 = 0;
+    mat.m42 = 0;
+
+    // guard against non-invertible matrices
+    let localDx = dx;
+    let localDy = dy;
+    try {
+        const inv = mat.inverse();
+        const localDelta = new DOMPoint(dx, dy).matrixTransform(inv);
+        localDx = localDelta.x;
+        localDy = localDelta.y;
+    } catch (err) {
+        // fallback: if inverse fails, use raw screen delta
+    }
+
+    layer.transform.matrix = layer.transform.matrix.translate(localDx, localDy);
+}
